posts/serializer.py
from dataclasses import fields
from xmlrpc.client import DateTime
from rest_framework import serializers
from .models import Post


class PostSerializers(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['id', 'title', 'content', 'created']
    