requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==4.1.2
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.5
